requirements.txt
Django==5.0.6
python-decouple==3.8
pyserial==3.5
requests==2.31.0
django-crispy-forms==2.1
crispy-bootstrap4==2023.1
twilio==8.12.0
razorpay==1.4.1